refactor: more standaloning of ops metrics
